requirements.txt
# PDF parsing libraries
PyMuPDF==1.26.5          # Python 3.9 compatible version
pdfplumber>=0.10.0       # Tested on Python 3.9
pypdf>=3.0.0             # Pure Python, 3.9 compatible

# Test data generation
reportlab>=4.0.0         # For generating synthetic PDFs

# Benchmarking and output
tabulate>=0.9.0          # For formatted benchmark tables

# Configuration management
python-dotenv>=1.0.0     # Environment variable management, Python 3.9 compatible
